+++ typespec-extension/changelog.md
@@ -4,6 +4,8 @@
 
 Compatible with compiler 0.58.
 
+- Supported `@multipartBody` from typespec http lib.
+
 ## 0.18.2 (2024-07-22)
 
 Compatible with compiler 0.58.